Generate chunks on server, send to client

+++ client/net/Connection.ts
@@ -2,7 +2,7 @@ const MAX_ATTEMPTS = 3
 
 type ConnectionState =
   | { type: 'ws'; ws: WebSocket; error: boolean; attemptsLeft: number }
-  | { type: 'worker'; worker: Worker }
+  | { type: 'worker'; worker: Worker | Window }
   | null
 
 export class Connection<ReceiveType, SendType = never> {
@@ -43,14 +43,21 @@ export class Connection<ReceiveType, SendType = never> {
     this.#state = { type: 'ws', ws, error: false, attemptsLeft }
   }
 
-  connectWorker (path: string): void {
-    const worker = new Worker(path)
+  connectWorker (path?: string): void {
+    const worker = path ? new Worker(path) : self
     for (const message of this.#queue) {
       worker.postMessage(message)
     }
-    worker.addEventListener('message', e => {
-      this.handleMessage(e.data)
-    })
+    // TypeScript hack; for some reason these cases can't merge
+    if (worker instanceof Worker) {
+      worker.addEventListener('message', e => {
+        this.handleMessage(e.data)
+      })
+    } else {
+      worker.addEventListener('message', e => {
+        this.handleMessage(e.data)
+      })
+    }
     this.#state = { type: 'worker', worker }
   }
 
@@ -63,12 +70,14 @@ export class Connection<ReceiveType, SendType = never> {
         this.#state.ws.close()
       }
     } else if (this.#state?.type === 'worker') {
-      this.#state.worker.terminate()
+      if (this.#state.worker instanceof Worker) {
+        this.#state.worker.terminate()
+      }
     }
     this.#state = null
   }
 
-  send (message: SendType): void {
+  send (message: SendType, transfer: Transferable[] = []): void {
     if (
       this.#state?.type === 'ws' &&
       this.#state.ws.readyState === WebSocket.OPEN
@@ -76,7 +85,7 @@ export class Connection<ReceiveType, SendType = never> {
       // TODO: Binary data isn't JSON-serializable this way
       this.#state.ws.send(JSON.stringify(message))
     } else if (this.#state?.type === 'worker') {
-      this.#state.worker.postMessage(message)
+      this.#state.worker.postMessage(message, { transfer })
     } else {
       this.#queue.push(message)
     }
